Implement full snake body.
Implement a snake body using a contained single-linked list structure, where each node in the list points to the next node and is contained inside of a struct that represents a snake body segment.
The body is expanded with another segment when a food item is eaten by the snake.
The game ends if the snake head collides with any of the snake body segments.